requirements.txt
allure-behave==2.13.5
allure-pytest==2.13.5
behave==1.2.6
black==24.3.0
faker==24.8.0
httpx==0.27.0
importlib-metadata==7.1.0
more-itertools==10.2.0
openpyxl==3.1.2
pre-commit==3.7.0
py==1.11.0
pylint==3.1.0
pymsteams==0.2.2
pyparsing==3.1.2
pytest==8.1.1
pytest-excel==1.6.0
pytest-html==4.1.1
pytest-md-report==0.5.1
pytest-retry==1.6.2
pytest-runner==6.0.1
pytest-selenium==4.1.0
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
selenium==4.19.0
wcwidth==0.1.8
webdriver-manager==4.0.1
zipp==3.0.0
